+++ frontend/src/components/Common/Description/index.ts
@@ -1,0 +1,11 @@
+/*!
+ * This file is part of the Stooa codebase.
+ *
+ * (c) 2020 - present Runroom SL
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+export * from './Description';
+export { default } from './Description';
