+++ frontend/src/landing/HomeSections/Benefits.tsx
@@ -12,7 +12,8 @@ import useTranslation from 'next-translate/useTranslation';
 import lottie from 'lottie-web';
 
 import { Lottie } from '@/types/animations';
-import { Column, Description, Row, Wrapper } from '@/landing/ui/styles';
+import { Column, Row, Wrapper } from '@/landing/ui/styles';
+import Description from '@/components/Common/Description';
 
 interface Props {
   item: Lottie;
