+++ frontend/src/pages/recover-password.tsx
@@ -12,7 +12,7 @@ import useTranslation from 'next-translate/useTranslation';
 import { useAuth } from '@/contexts/AuthContext';
 import RecoverPasswordForm from '@/components/Web/Forms/recover-password';
 import Layout from '@/layouts/Default';
-import { Description } from '@/ui/pages';
+import Description from '@/components/Common/Description';
 
 const RecoverPassword = () => {
   const { t } = useTranslation('recover');
